Modularizamos en un controlador la logica de los comandos y mejoramos algunos errores de sintaxis
