+++ src/index.css
@@ -5,5 +5,4 @@
 
 .contenedorTareas {
   background-color: rgba(0, 0, 0, 0.472);
-  margin: 20px 20px 20px 20px;
 }
